move typescript parser code
