+++ src/CodeLensProvider.ts
@@ -1,5 +1,4 @@
 import * as vscode from 'vscode';
-import ts from 'typescript';
 import { logger, startDebuggingCodeLensCommandName } from './extension';
 import { Pool, PoolClient } from 'pg';
 import { DbosConfig, loadConfigFile, locateDbosConfigFile } from './dbosConfig';
@@ -8,6 +7,7 @@ import { DbosCloudApp, getApp, getDbCredentials, getDbInstance, isUnauthorized }
 import path from 'node:path';
 import { DebugProxyManager } from './DebugProxyManager';
 import { Configuration } from './Configuration';
+import { parseTypeScript } from './parsers/typeScript';
 
 interface workflow_status {
     workflow_uuid: string;
@@ -417,132 +417,9 @@ export class CodeLensProvider implements vscode.CodeLensProvider<DbosCodeLens>, 
     }
 }
 
-export interface NamedImportInfo {
-    name: string;
-    alias: string;
-    moduleName: string;
-}
-
-export interface DecoratorInfo {
-    name: string;
-    propertyName?: string;
-}
-
-export interface StaticMethodInfo {
-    name: string;
-    className: string | undefined;
-    start: number;
-    end: number;
-    decorators: DecoratorInfo[];
-}
-
-interface DbosWorkflowMethod {
+export interface DbosWorkflowMethod {
     name: string;
     start: vscode.Position;
     end: vscode.Position;
 }
 
-function* parseTypeScript(document: vscode.TextDocument, token: vscode.CancellationToken): Generator<DbosWorkflowMethod, void, unknown> {
-    const file = ts.createSourceFile(
-        document.fileName,
-        document.getText(),
-        ts.ScriptTarget.Latest
-    );
-
-    for (const method of getWorkflowMethods(file)) {
-        if (token.isCancellationRequested) { break; }
-        yield { 
-            start: document.positionAt(method.start),
-            end: document.positionAt(method.end),
-            name: method.name,
-        };
-    }
-}
-
-export function* getWorkflowMethods(file: ts.SourceFile) {
-    const importMap = new Map<string, NamedImportInfo>();
-    for (const imp of getImports(file)) {
-        importMap.set(imp.alias, imp);
-    }
-
-    for (const method of getStaticMethods(file)) {
-        for (const d of method.decorators) {
-            const imp = importMap.get(d.name);
-            if (!imp) { continue; }
-            if (imp.moduleName !== '@dbos-inc/dbos-sdk') { continue; }
-            if (imp.name === 'Workflow' && d.propertyName === undefined) {
-                yield method;
-            }
-            else if (imp.name === 'DBOS' && d.propertyName === 'workflow') {
-                yield method;
-            }
-        }
-    }
-}
-
-export function* getImports(file: ts.SourceFile): Generator<NamedImportInfo, void, unknown> {
-    for (const node of file.statements) {
-        if (ts.isImportDeclaration(node)) {
-            const moduleName = getName(node.moduleSpecifier);
-            const bindings = node.importClause?.namedBindings;
-            if (!bindings) { continue; }
-            else if (ts.isNamedImports(bindings)) {
-                for (const binding of bindings.elements) {
-                    const name = binding.propertyName?.text ?? binding.name.text;
-                    const alias = binding.name.text;
-                    yield { name, alias, moduleName };
-                }
-            }
-            else {
-                throw Error(`Unsupported NamedImportBindings kind: ${ts.SyntaxKind[node.kind]}`);
-            }
-        }
-    }
-}
-
-export function* getStaticMethods(file: ts.SourceFile): Generator<StaticMethodInfo, void, unknown> {
-    for (const node of file.statements) {
-        if (ts.isClassDeclaration(node)) {
-            const className = node.name?.text;
-            for (const memberNode of node.members) {
-                if (!ts.isMethodDeclaration(memberNode)) { continue; }
-
-                const isStatic = (memberNode.modifiers ?? []).some(m => m.kind === ts.SyntaxKind.StaticKeyword);
-                if (!isStatic) { continue; }
-
-                const decorators = (ts.getDecorators(memberNode) ?? [])
-                    .map(parseDecorator)
-                    .filter(isValid);
-                yield {
-                    name: getName(memberNode.name),
-                    className,
-                    start: memberNode.getStart(file),
-                    end: memberNode.getEnd(),
-                    decorators,
-                };
-            }
-        }
-    }
-}
-
-export function parseDecorator(node: ts.Decorator): DecoratorInfo | undefined {
-    if (!ts.isCallExpression(node.expression)) { return; }
-    const expr = node.expression.expression;
-    if (ts.isIdentifier(expr)) {
-        return { name: expr.text, propertyName: undefined };
-    }
-    if (ts.isPropertyAccessExpression(expr)) {
-        return { name: getName(expr.expression), propertyName: expr.name.text };
-    }
-}
-
-function isValid<T>(value: T | null | undefined): value is T { return !!value; }
-
-function getName(node: ts.PropertyName | ts.Expression | ts.LeftHandSideExpression) {
-    switch (true) {
-        case ts.isCallExpression(node): return getName(node.expression);
-        case ts.isIdentifier(node): return node.text;
-        case ts.isStringLiteral(node): return node.text;
-        default: throw Error(`Unsupported name kind: ${ts.SyntaxKind[node.kind]}`);
-    }
-}
